checkpoint with hanging requests to openam
