+++ src/main/scala/org/broadinstitute/dsde/workbench/sam/service/ResourceService.scala
@@ -1,7 +1,7 @@
 package org.broadinstitute.dsde.workbench.sam.service
 
 import akka.http.scaladsl.model.{StatusCode, StatusCodes}
-import org.broadinstitute.dsde.workbench.sam.dataaccess.{AccessManagementDAO, DirectoryDAO}
+import org.broadinstitute.dsde.workbench.sam.dataaccess.{OpenAmDAO, DirectoryDAO}
 import org.broadinstitute.dsde.workbench.sam.model.SamModels._
 
 import scala.concurrent.{ExecutionContext, Future}
@@ -9,18 +9,22 @@ import scala.concurrent.{ExecutionContext, Future}
 /**
   * Created by mbemis on 5/22/17.
   */
-class ResourceService(val accessManagementDAO: AccessManagementDAO, val directoryDAO: DirectoryDAO)(implicit val executionContext: ExecutionContext) {
+class ResourceService(val openAmDAO: OpenAmDAO, val directoryDAO: DirectoryDAO)(implicit val executionContext: ExecutionContext) {
 
   def createResourceType(resource: Resource): Future[Boolean] = {
     println(s"Creating resource: $resource")
 
-    //Create the resource type if it doesn't exist
-    //TODO
+//    //Create the resource type if it doesn't exist
+//    openAmDAO.createResourceType(resource)
+//
+//    //Create the roles and set their actions for the resource type
+//    Future.traverse(resource.roles) { createResourceRole(resource.resourceType, _) }
 
-    //Create the roles and set their actions for the resource type
-    Future.traverse(resource.roles) { createResourceRole(resource.resourceType, _) }
 
-    Future.successful(true)
+    for {
+      _ <- openAmDAO.createResourceType(resource)
+      _ <- Future.traverse(resource.roles) { createResourceRole(resource.resourceType, _) }
+    } yield true
   }
 
   def createResourceRole(resourceType: String, role: ResourceRole): Future[Boolean] = {
